+++ CHANGES.md
@@ -3,6 +3,14 @@ Change Log
 
 Beta Releases
 -------------
+
+### b29 - 2014-06-02
+
+* Breaking changes
+  * Removed `CesiumWidget.onRenderLoopError` and `Viewer.renderLoopError`.  They have been replaced by `Scene.renderError`.
+* Improved terrain and imagery rendering performance when very close to the surface.
+* Added `preRender` and `postRender` events to `Scene`.
+
 ### b28 - 2014-05-01
 
 * Breaking changes ([why so many?](https://groups.google.com/forum/#!topic/cesium-dev/CQ0wCHjJ9x4)):